[fix](s3) move header file
