+++ cloud/test/s3_accessor_mock_test.cpp
@@ -31,7 +31,7 @@
 #include "common/config.h"
 #include "common/logging.h"
 #include "cpp/aws_common.h"
-#include "cpp/container_credentials_test_util.h"
+#include "cpp/test/container_credentials_test_util.h"
 #include "cpp/obj-client/s3_obj_storage_client.h"
 #include "cpp/sync_point.h"
 #include "recycler/s3_accessor.h"
